counterpick-app/apps/backend/requirements.txt
requests
lxml
pytest
urllib3
flask
flask-cors
flask-socketio>=5.3.0
python-socketio>=5.9.0
websocket-client>=1.6.0
supabase>=2.4.0
python-dotenv
